refactor: EditPage 스타일 코드 파일 분리 #21
