+++ frontend/src/pages/EditPage/EditPage.styled.ts
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+import { Tab } from '@/components/Tab';
+
+export const TabList = styled(Tab.List)`
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  height: 4.8rem;
+
+  button {
+    ${({ theme }) => theme.fonts.body15};
+
+    color: ${({ theme }) => theme.colors.gray400};
+
+    &[data-state='active'] {
+      color: ${({ theme }) => theme.colors.green700};
+    }
+  }
+`;
